+++ src/output/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: 2021-2023 Birger Schacht <birger@rantanplan.org>
+//
+// SPDX-License-Identifier: MIT
+
+pub mod agenda;
+pub mod calendar;
+pub mod date;
