+++ src/utils/mod.rs
@@ -3,6 +3,9 @@
 // SPDX-License-Identifier: MIT
 
 mod date_extensions;
-pub mod types;
+mod helpers;
+mod types;
 
 pub use date_extensions::{DateExtensions, MonthFullWeeksIter};
+pub use helpers::convertstyle;
+pub use types::{ChronoDate, ChronoDateTime};
